implement the Driver class
This moves public definitions out of protocol into the main
namespace, and out of Protocol.hpp into Types.hpp. Protocol.hpp
should be considered private, and should not be included in
Driver.hpp/Types.hpp anymore.

Main queries have been implemented and tested. This prepares the field
for a scheme for handling timestamping that either relies on the
device's own internal time (using e.g. GPS sync) or if not available the
host computer's. In the latter case, we setup the UnixTime packet to be
sent at each cycle.