implement periodic packet processing
This implementation tries to keep the ability to pick parts of the
updates at different periods in order to not overload the communication
channel.

The API is basically made of set.*Period methods that allow to pick
which data to update when, and the poll() method. See the poll method
documentation for more information on how it is supposed to work.