invalidate parts of the aggregate structures when the corresponding messages are disabled
While it's easy for the caller to know that it should not read GPS
data if the GPS message is disabled, it's a lot harder to know
that the orientation of mWorld is invalid but the other fields are
valid.

Since we do the aggregation in-class, we should keep track of these
and invalidate the relevant fields.